(feat) Добавлен перенос карт в руке во входные нейроны сети

+++ src/ru/gigabyte_artur/blackjack_ai/Neuron.java
@@ -16,7 +16,7 @@ public class Neuron
     }
 
     // Устанавливает сигнал нейрона в signal_in.
-    public void SetSignal(int signal_in)
+    public void SetSignal(double signal_in)
     {
         this.signal = signal_in;
     }
